restored exception handling
