+++ bin/acli
@@ -17,10 +17,15 @@
 namespace Acquia\Cli;
 
 use Acquia\Cli\Command\Api\ApiCommandHelper;
+use Acquia\Cli\Exception\AcquiaCliException;
+use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
 use Symfony\Component\Console\Application;
+use Symfony\Component\Console\ConsoleEvents;
+use Symfony\Component\Console\Event\ConsoleErrorEvent;
 use Symfony\Component\Console\Input\ArgvInput;
 use Symfony\Component\Dotenv\Dotenv;
 use Symfony\Component\ErrorHandler\Debug;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\Filesystem\Filesystem;
 
 if (!in_array(PHP_SAPI, ['cli', 'phpdbg', 'embed'], true)) {
@@ -73,5 +78,16 @@ $application->setVersion('@package_version@');
 /** @var \Acquia\Cli\Command\Api\ApiCommandHelper $helper */
 $helper = $container->get(ApiCommandHelper::class);
 $application->addCommands($helper->getApiCommands());
+// Clean up exceptions thrown during commands.
+$dispatcher = new EventDispatcher();
+$dispatcher->addListener(ConsoleEvents::ERROR, function (ConsoleErrorEvent $event) {
+  $exitCode = $event->getExitCode();
+  $error = $event->getError();
+  // Make OAuth server errors more human-friendly.
+  if ($error instanceof IdentityProviderException && $error->getMessage() === 'invalid_client') {
+    $event->setError(new AcquiaCliException('Your Cloud API credentials are invalid. Run acli auth:login to reset them.', [], $exitCode));
+  }
+});
+$application->setDispatcher($dispatcher);
 
 $application->run();
