+++ tests/phpunit/src/TestBase.php
@@ -2,6 +2,7 @@
 
 namespace Acquia\Cli\Tests;
 
+use Acquia\Cli\Exception\AcquiaCliException;
 use Acquia\Cli\Helpers\ClientService;
 use Acquia\Cli\Helpers\DataStoreContract;
 use Acquia\Cli\Helpers\LocalMachineHelper;
@@ -9,6 +10,7 @@ use Acquia\Cli\Helpers\SshHelper;
 use Acquia\Cli\Helpers\TelemetryHelper;
 use AcquiaCloudApi\Connector\Client;
 use AcquiaLogstream\LogstreamManager;
+use League\OAuth2\Client\Provider\Exception\IdentityProviderException;
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
 use Prophecy\Prophet;
@@ -17,10 +19,14 @@ use Symfony\Component\Cache\Adapter\FilesystemAdapter;
 use Symfony\Component\Cache\Adapter\PhpArrayAdapter;
 use Symfony\Component\Cache\CacheItem;
 use Symfony\Component\Console\Application;
+use Symfony\Component\Console\Command\Command;
+use Symfony\Component\Console\ConsoleEvents;
+use Symfony\Component\Console\Event\ConsoleErrorEvent;
 use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Logger\ConsoleLogger;
 use Symfony\Component\Console\Output\BufferedOutput;
 use Symfony\Component\Console\Output\ConsoleOutput;
+use Symfony\Component\EventDispatcher\EventDispatcher;
 use Symfony\Component\Filesystem\Filesystem;
 use Symfony\Component\Finder\Finder;
 use Symfony\Component\Yaml\Yaml;
@@ -177,6 +183,18 @@ abstract class TestBase extends TestCase {
     $this->logStreamManagerProphecy = $this->prophet->prophesize(LogstreamManager::class);
     $this->sshHelper = new SshHelper($output, $this->localMachineHelper);
 
+    // Clean up exceptions thrown during commands.
+    $dispatcher = new EventDispatcher();
+    $dispatcher->addListener(ConsoleEvents::ERROR, function (ConsoleErrorEvent $event) {
+      $exitCode = $event->getExitCode();
+      $error = $event->getError();
+      // Make OAuth server errors more human-friendly.
+      if ($error instanceof IdentityProviderException && $error->getMessage() === 'invalid_client') {
+        $event->setError(new AcquiaCliException('Your Cloud API credentials are invalid. Run acli auth:login to reset them.', [], $exitCode));
+      }
+    });
+    $this->application->setDispatcher($dispatcher);
+
     $this->removeMockConfigFiles();
     $this->createMockConfigFile();
 
@@ -235,6 +253,35 @@ abstract class TestBase extends TestCase {
     }
 
     return json_decode($response_body);
+  }
+
+  /**
+   * Build and return a command with common dependencies.
+   *
+   * All commands inherit from a common base and use the same constructor with a
+   * bunch of dependencies injected. It would be tedious for every command test
+   * to inject every dependency as part of createCommand(). They can use this
+   * instead.
+   *
+   * @param string $commandName
+   *
+   * @return \Symfony\Component\Console\Command\Command
+   */
+  protected function injectCommand(string $commandName): Command {
+    return new $commandName(
+      $this->cloudConfigFilepath,
+      $this->localMachineHelper,
+      $this->cloudDatastore,
+      $this->acliDatastore,
+      $this->telemetryHelper,
+      $this->amplitudeProphecy->reveal(),
+      $this->acliConfigFilename,
+      $this->projectFixtureDir,
+      $this->clientServiceProphecy->reveal(),
+      $this->logStreamManagerProphecy->reveal(),
+      $this->sshHelper,
+      $this->sshDir
+    );
   }
 
   /**
